# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NOVA Ride API",
    "version": "1.0.0",
    "description": "جرد مسارات مُولّد آليًا من المتحكّمات. البادئة /api؛ كل المسارات متاحة أيضًا تحت /api/v1. لا يتضمّن هذا الجرد مخطّطات الطلب/الاستجابة (تُثرى لاحقًا)."
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:4000"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "auth_logout",
        "summary": "POST /api/auth/logout",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/auth/me": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "auth_me",
        "summary": "POST /api/auth/me",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/backups": {
      "get": {
        "tags": [
          "backups"
        ],
        "operationId": "backups_list",
        "summary": "GET /api/backups",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      },
      "post": {
        "tags": [
          "backups"
        ],
        "operationId": "backups_create",
        "summary": "POST /api/backups",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      }
    },
    "/api/backups/:id": {
      "delete": {
        "tags": [
          "backups"
        ],
        "operationId": "backups_remove",
        "summary": "DELETE /api/backups/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      },
      "get": {
        "tags": [
          "backups"
        ],
        "operationId": "backups_findOne",
        "summary": "GET /api/backups/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      },
      "patch": {
        "tags": [
          "backups"
        ],
        "operationId": "backups_update",
        "summary": "PATCH /api/backups/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      }
    },
    "/api/backups/dr-status": {
      "get": {
        "tags": [
          "backups"
        ],
        "operationId": "backups_drStatus",
        "summary": "GET /api/backups/dr-status",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      }
    },
    "/api/backups/retention/apply": {
      "post": {
        "tags": [
          "backups"
        ],
        "operationId": "backups_applyRetention",
        "summary": "POST /api/backups/retention/apply",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      }
    },
...
    "/api/content-blocks": {
      "get": {
        "tags": [
          "content-blocks"
        ],
        "operationId": "content-blocks_list",
        "summary": "GET /api/content-blocks",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      },
      "post": {
        "tags": [
          "content-blocks"
        ],
        "operationId": "content-blocks_create",
        "summary": "POST /api/content-blocks",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      }
    },
    "/api/content-blocks/:id": {
      "delete": {
        "tags": [
          "content-blocks"
        ],
        "operationId": "content-blocks_remove",
        "summary": "DELETE /api/content-blocks/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      },
      "get": {
        "tags": [
          "content-blocks"
        ],
        "operationId": "content-blocks_findOne",
        "summary": "GET /api/content-blocks/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      },
      "patch": {
        "tags": [
          "content-blocks"
        ],
        "operationId": "content-blocks_update",
        "summary": "PATCH /api/content-blocks/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: settings.manage"
      }
    },
...
    "/api/driver/me": {
      "get": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_me",
        "summary": "GET /api/driver/me",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_updateMe",
        "summary": "PATCH /api/driver/me",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/driver/me/availability": {
      "post": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_availability",
        "summary": "POST /api/driver/me/availability",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/driver/me/documents": {
      "post": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_addDocument",
        "summary": "POST /api/driver/me/documents",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/driver/me/earnings": {
      "get": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_earnings",
        "summary": "GET /api/driver/me/earnings",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/driver/me/sanctions": {
      "get": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_sanctions",
        "summary": "GET /api/driver/me/sanctions",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/driver/me/trips": {
      "get": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_trips",
        "summary": "GET /api/driver/me/trips",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/driver/me/trips/:id": {
      "get": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_trip",
        "summary": "GET /api/driver/me/trips/:id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/driver/me/trips/:id/status": {
      "patch": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_updateTrip",
        "summary": "PATCH /api/driver/me/trips/:id/status",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/driver/me/upload-url": {
      "post": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_uploadUrl",
        "summary": "POST /api/driver/me/upload-url",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/drivers": {
      "get": {
        "tags": [
          "drivers"
        ],
        "operationId": "drivers_findAll",
        "summary": "GET /api/drivers",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "الصلاحيات المطلوبة: drivers.read, drivers.manage"
      }
    },
...
  }
}